# Issue #1355 Slight optimization for performance regrid array