--- a/imod/mf6/utilities/regrid.py
+++ b/imod/mf6/utilities/regrid.py
@@ -161,8 +161,8 @@ def _regrid_array(
     - None
     """
 
-    # skip regridding for arrays with no valid values (such as "None")
-    if not _is_valid(da.values[()]):
+    # skip regridding for scalar arrays with no valid values (such as "None")
+    if is_scalar(da) and not _is_valid(da.values[()]):
         return None
 
     # the dataarray might be a scalar. If it is, then it does not need regridding.