--- a/imod/mf6/utilities/regrid.py
+++ b/imod/mf6/utilities/regrid.py
@@ -162,11 +162,12 @@ def _regrid_array(
     """
 
     # skip regridding for scalar arrays with no valid values (such as "None")
-    if is_scalar(da) and not _is_valid(da.values[()]):
+    scalar_da: bool = is_scalar(da)
+    if scalar_da and not _is_valid(da.values[()]):
         return None
 
     # the dataarray might be a scalar. If it is, then it does not need regridding.
-    if is_scalar(da):
+    if scalar_da:
         return da.values[()]  # type: ignore [attr-defined]
 
     if isinstance(da, xr.DataArray):