--- a/imod/mf6/utilities/regrid.py
+++ b/imod/mf6/utilities/regrid.py
@@ -168,7 +168,7 @@ def _regrid_array(
 
     # the dataarray might be a scalar. If it is, then it does not need regridding.
     if scalar_da:
-        return da.values[()]  # type: ignore [attr-defined]
+        return da.values[()]
 
     if isinstance(da, xr.DataArray):
         coords = da.coords