Fixes #1355

# Description
Slight performance improvement: Avoid checking if all arrays is_invalid. Only do when ``is_scalar``.

For test with 1e9 cells:

Before fix:
567 ms ± 26.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
After fix:
545 ms ± 19.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Second time I ran these tests:
Before fix:
530 ms ± 21.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
After fix:
503 ms ± 29.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Script to run:

```python
import imod
import numpy as np
from imod.mf6.utilities.regrid import _regrid_array, RegridderWeightsCache, RegridderType
import xarray as xr

dx = 1.
dy = -1.
xmin = 0.0
xmax = 10000.0
ymin = 0.0
ymax = 10000.0
layer = np.arange(10) + 1

da = xr.ones_like(imod.util.empty_3d(dx=dx, xmin=xmin, xmax=xmax, dy=dy, ymin=ymin, ymax=ymax, layer=layer))
target = imod.util.empty_2d(dx=10.0, xmin=xmin, xmax=xmax, dy=-10.0, ymin=ymin, ymax=ymax)

regridder_collection = RegridderWeightsCache()
regridder_name = RegridderType.OVERLAP 
regridder_function = "mean"

# In IPython, run:
# %timeit _regrid_array(da, regridder_collection, regridder_name, regridder_function, target)
```


# Checklist

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example