@JoerivanEngelen 2024-12-19
> Would be interesting to see how well this scales for bigger datagrids

I don't foresee usecases for larger grids than 1e9 cells :P But I foresee use cases where a LOT of these grids have to be regridded. Regardless, it is going to at least slightly improve performance. This was some low-hanging fruit.